Fix: Updated bloc listeners to unsubscribe from bloc stream when widget is disposed.
Achieved it by replacing useMemoized with useEffect. I believe it was actually intented to use useEffect there, because function inside of it returns subscription cancellation function, but because useMemoized was used, it was never called. So, it caused listener to be executed more than once in some cases. For example, when listener would be attached inside of dialog which we would open/close multiple times.